force index cast in case 32 bit int

+++ flash_attn/flash_attn_triton.py
@@ -190,7 +190,7 @@ def _fwd_kernel(
                 (start_n + offs_n)[None, :]
 
             # attn matrix has shape (B * H, seqlen_q, seqlen_k)
-            rand = tl.rand(dropout_seed, dropout_seq_offset + indices)
+            rand = tl.rand(dropout_seed, dropout_seq_offset.to(indices.dtype) + indices)
 
             p *= tl.where(rand > dropout_p, 1.0 / (1.0 - dropout_p), 0.0)
 
@@ -421,7 +421,7 @@ def _bwd_kernel_one_col_block(
                 (offs_m_curr * seqlen_k)[:, None] + \
                 offs_n[None, :]
 
-            rand = tl.rand(dropout_seed, dropout_seq_offset + indices)
+            rand = tl.rand(dropout_seed, dropout_seq_offset.to(indices.dtype) + indices)
             keep_mask = rand > dropout_p
             zij = keep_mask.to(tl.float32) * (1.0 / (1.0 - dropout_p))
